Windows CI: test-unit pkg\archive step 1
